xe_nano: Add Hardware counter class
Create a hardware counter class that can be used to retrieve
hardware counters such as instruction and cycle counts.

This class is dependent on the PAPI library to access the
hardware counters.  If the library is not available in the
system, cmake will issue a warning.

INTERNAL: VLCLJ-93 - Implement Instruction and cycle counts for
INTERNAL: nano benchmark

Change-Id: I86ca2140da6286a8784340514e62194cc65236ec
Signed-off-by: Sebastian Sanchez <sebastian.sanchez@intel.com>