Add xello_interop_shared_compute black-box test.
This test uses CL and XE to compute a matrix multiply, with each writing
alternating elements of the result matrix.  This verifies that memory
buffer sharing works as expected.

The common code is also generalized and refactored a bit to separate
compute from memory copy in and copy out, and to read in the .cl program
source from a file rather than have it inline.

Change-Id: I6d661a0de46223ac13e6fc41f3f3ad17100c9ebb
Signed-off-by: Jim Snow <jim.m.snow@intel.com>