Fix infinite loop on null driver initialization
Null driver is assigning the base xeInit to the ddi table, instead
of that one defined for the null driver (probably because of having
the same name and the linker assigning the first one it finds).

So define the functions for null driver as *_null to break the
infinite loop of calling xeInit->loading loader->filling table.

Signed-off-by: Jaime Arteaga <jaime.a.arteaga.molina@intel.com>