Changes to scripting framework:
(1) Enable typedef to known types - fix bug where the type isn't changed to have the appropriate namespace/class prefix.
(2) to_string of structures containing array elements to known types now outputted correctly - resolves compiler errors due to no known conversion when the member of the structure is an array.
(3) Permit structure members to have array sizes that are defined in an enum. Without this fix, the array size is copied verbatum and references the C definition which is not known - need to point to the appropate enum class that contains the size enum string.