debug: shared vs. mirrored module heap
For shared module heaps we only allow attaching to the device in order to
avoid leaking breakpoints to other debug sessions.

For mirrored module heaps we allow attaching either to the device or to
sub-devices.

Debuggers that support both modes may probe this by attempting to attach
to sub-devices and fall back to attaching to the parent device.

Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>