debug, memory: memory spaces
Adding a memory space parameter to memory APIs based on a request from Ben
Ashbaugh in the 2019/10/24 review.

The memory space selector is declared as an int and supported by a
device-specific enumeration requested by Ben Ashbaugh in the 2019/11/14
review.

The current thinking is that compilers would specify the memory space via
the DW_AT_segment attribute.  This may be subject to change but will not
affect level-zero.

Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>