debug, event: module unload
Add a module unload event symmetric to module load.  The event is sent in
the context of the $xModuleDestroy() flow before the module is actually
unloaded.  The debugger needs to acknowledge the event before UMD can
unload the module and reclaim the memory.

The event has been discussed with Jaroslaw Chodor.

Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>