debug, event: remove loadEnd field from module events
Providing a virtual address range for the loaded module would not scale to
modules occupying a non-contiguous address range.  While we do not need
this flexibility today, we also do not need to restrict it.

Debuggers need to know the load address in order to relocate debug
information.  The layout of the module is described in the ELF program
header.  In theory, we could allow gaps between segments.

Remove the loadEnd field and rename loadBegin to just load.

Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>