debug, event: remove exception payload
The exception event provides a few registers typically required by every
debugger:

  * the exception code
  * the instruction pointer
  * the stack pointer

This is an optimization to avoid a round-trip for reading those registers
in some cases.

For GEN, we would further like to provide emask, although this won't have
an equivalent for non-SIMD architectures.

For other architectures, we may want to provide additional or different
registers.

It is further not clear how much this round-trip actually costs in the
overall flow.

Remove the payload fields.  If it turns out to be useful enough to have
such a feature, we can add it or something similar in a future version.

Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>