debug, run-control: do not allow interrupting unavailable threads
We allow interrupting a thread that had been stopped at an internal
breakpoint on kernel entry in order to prevent it from entering the kernel
when the debug session has been stopped.

This would allow subsequent calls to zetDebugInterrupt() for the same
thread to produce different results if that thread hit the entry
breakpoint between those calls while the debug session is stopped.

Remove that special case and report threads stopped at entry breakpoints
as unavailable consistently.

Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>