tools, debug: fix register read/write interface
This fixes the debug tools register state interface.  The interface has not been
implemented, yet, and there are no plans to implement the old version.

I did not guard the changes by a future version.  If necessary, this can be done
but it would complicate things for users, for NEO, and in the spec.

Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>