Update ze_device_properties_t & add extension for numTotalEUs
This commit updates the description of the numThreadsPerEU, numEUsPerSubslice,
numSubslicesPerSlice, and numSlices attributes of the ze_device_properties_t struct
to specify that the returned values are the maximum expected. It also adds a new
standard extension “ZE_extension_capability_query” that allows applications to query
the total number of EUs available in the device.